added planetable.pl utility, fixed some bugs
Fixed a bug in save_file that prevented it from using
the stored file.

Added missing get() to documentation synopsis.

Added utility planetable.pl for building a table
from a function that defines a plane.
This can be used after a least squares fit to
quickly define an entire table.