added lookup_points(), fixed order, etc
Added lookup_points() for lookup up points that are
relative to give values.

Due to some irregularities the position orientation of
the tables had to be revised.
In order for lookup_points to work the columns coordinates
must be in increasing order.
Previously the vertical coordiantes were reversed.

Removed get_*_vals, flatten.